requirements.txt
# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pillow>=9.0.0

# Machine Learning & Deep Learning
torch>=1.11.0
torchvision>=0.12.0
scikit-learn>=1.0.0
opencv-python>=4.5.0

# GUI Frameworks
tkinter-dev  # Usually comes with Python
PyQt5>=5.15.0
customtkinter>=5.0.0

# Database
sqlite3  # Usually comes with Python
sqlalchemy>=1.4.0

# Image Processing & Computer Vision
imageio>=2.15.0
scikit-image>=0.19.0

# Data Handling
pandas>=1.4.0
h5py>=3.6.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
click>=8.0.0
colorama>=0.4.4

# Development & Testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: For enhanced GUI
plotly>=5.0.0
dash>=2.0.0  # For web-based interface alternative

# Optional: For advanced ML features
transformers>=4.15.0  # If we want to experiment with vision transformers
tensorboard>=2.8.0  # For training visualization
